+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- All actions are now dumb aware
 - Display mode settings are now grouped into tabs
 
 ## [1.0.2] - 2024-03-29